refactor: Update database URL in .env.example
Update the database URL in the .env.example file to use the "vidify" database instead of "clearvault".

Deleted migrations for "chat" and "videos" tables

Remove the migrations for the "chat" and "videos" tables as they are no longer needed.

Update "compose" script in package.json

Update the "compose" script in the package.json file to use the "docker compose" command instead of "docker-compose".

Update console log message in app.ts

Update the console log message in the app.ts file to display the correct port number.

Update console log messages in socket.ts

Update the console log messages in the socket.ts file to display the correct log messages.

Add migrations for "videos" and "chat" tables

Add migrations for the "videos" and "chat" tables to create the necessary database tables.